Introduce struct for holding stream info
